refactor admin pages with shared components
